major spotify changes
Added navigation, new dmx light, spotify login and flow control
