+++ .gitignore
@@ -10,3 +10,7 @@
 *.bin
 .idea/workspace.xml
 app/app-release.apk
+*.bin
+.gradle/2.4/taskArtifacts/cache.properties.lock
+*.bin
+.idea/workspace.xml
