+++ .idea/libraries/spotify_player_1_0_0_beta10.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="spotify-player-1.0.0-beta10">
     <CLASSES>
+      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.spotify.sdk/spotify-player/1.0.0-beta10/jars/classes.jar!/" />
       <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.spotify.sdk/spotify-player/1.0.0-beta10/jars/libs/jnihelpers-1.0.jar!/" />
       <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.spotify.sdk/spotify-player/1.0.0-beta10/res" />
-      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.spotify.sdk/spotify-player/1.0.0-beta10/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
