+++ app/src/main/java/com/example/joe/mbls/spotify/SpotifyDisplayAlbums.java
@@ -1,14 +1,19 @@
 package com.example.joe.mbls.spotify;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.util.Log;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
 
+import com.MainMenu;
 import com.R;
 
 import java.util.ArrayList;
@@ -17,7 +22,7 @@ import kaaes.spotify.webapi.android.SpotifyService;
 import kaaes.spotify.webapi.android.models.Album;
 import kaaes.spotify.webapi.android.models.Pager;
 
-public class SpotifyDisplayAlbums extends Activity {
+public class SpotifyDisplayAlbums extends AppCompatActivity {
 
     private SpotifyService spotifyService;
     private ListView listView;
@@ -33,6 +38,15 @@ public class SpotifyDisplayAlbums extends Activity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_display_items_list);
 
+        Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+
+        setSupportActionBar(myToolbar);
+
+        // Get a support ActionBar corresponding to this toolbar
+        ActionBar ab = getSupportActionBar();
+        // Enable the Up button
+        ab.setDisplayHomeAsUpEnabled(true);
+
         SpotifyApplication spotifyApplication = ((SpotifyApplication) getApplicationContext());
         spotifyService = spotifyApplication.getSpotifyService();
 
@@ -45,11 +59,9 @@ public class SpotifyDisplayAlbums extends Activity {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 Album album = (Album) adapter.getItem(position);
-                Intent albumIntent = new Intent(getApplicationContext(), SpotifyDisplaySongs.class);
+                Intent albumIntent = new Intent(getApplicationContext(), SpotifyDisplayAlbumSongs.class);
 
                 albumIntent.putExtra("album", album.id);
-
-                Log.d("Albums", "calling display songs with album");
                 startActivity(albumIntent);
 
             }
@@ -59,7 +71,28 @@ public class SpotifyDisplayAlbums extends Activity {
 
     }
 
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.action_bar_spotify_main, menu);
+        return true;
+    }
 
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch(item.getItemId()) {
+            case R.id.menu_home:
+                Intent intent = new Intent(this, MainMenu.class);
+               // intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                startActivity(intent);
+                return true;
+            case R.id.spotify_go_home:
+                startActivity(new Intent(this, MainActivity.class));
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
 
     @Override
     protected void onDestroy() {
