+++ app/src/main/java/com/example/joe/mbls/spotify/SpotifyDisplayMySongs.java
@@ -1,10 +1,15 @@
 package com.example.joe.mbls.spotify;
 
 import android.app.Activity;
+import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.ContextMenu;
+import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
@@ -12,6 +17,7 @@ import android.widget.AdapterView;
 import android.widget.ImageButton;
 import android.widget.ListView;
 
+import com.MainMenu;
 import com.R;
 import com.spotify.sdk.android.player.Player;
 import com.spotify.sdk.android.player.PlayerNotificationCallback;
@@ -26,8 +32,7 @@ import kaaes.spotify.webapi.android.models.Pager;
 import kaaes.spotify.webapi.android.models.SavedTrack;
 import retrofit.http.QueryMap;
 
-public class SpotifyDisplayMySongs extends Activity implements
-        PlayerNotificationCallback {
+public class SpotifyDisplayMySongs extends AppCompatActivity {
 
     private SpotifyService spotifyService;
     private ListView listView;
@@ -47,6 +52,15 @@ public class SpotifyDisplayMySongs extends Activity implements
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_display_items_list);
 
+        Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+
+        setSupportActionBar(myToolbar);
+
+        // Get a support ActionBar corresponding to this toolbar
+        ActionBar ab = getSupportActionBar();
+        // Enable the Up button
+        ab.setDisplayHomeAsUpEnabled(true);
+
         SpotifyApplication spotifyApplication = ((SpotifyApplication) getApplicationContext());
         spotifyService = spotifyApplication.getSpotifyService();
         mPlayer = spotifyApplication.getPlayer();
@@ -62,6 +76,27 @@ public class SpotifyDisplayMySongs extends Activity implements
         new RetrieveMySongsTask().execute();
     }
 
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.action_bar_spotify_main, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch(item.getItemId()) {
+            case R.id.menu_home:
+                Intent intent = new Intent(this, MainMenu.class);
+                //intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                startActivity(intent);
+                return true;
+            case R.id.spotify_go_home:
+                startActivity(new Intent(this, MainActivity.class));
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
 
     @Override
     public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
@@ -77,8 +112,6 @@ public class SpotifyDisplayMySongs extends Activity implements
         switch(item.getItemId()) {
             case R.id.add_track_to_playlist:
                 mPlayer.queue(track.track.uri);
-                for (SavedTrack track1: tracks) Log.d("MYTRACKS BRO", track1.track.name);
-                Log.d("SpotifyDisplayMySongs", "Adding song to queue: " + track.track.id);
                 return true;
             case R.id.track_play_now:
                 mPlayer.clearQueue();
@@ -92,21 +125,7 @@ public class SpotifyDisplayMySongs extends Activity implements
 
 
 
-    @Override
-    public void onPlaybackEvent(EventType eventType, PlayerState playerState) {
-        Log.d("MainActivity", "Playback event received: " + eventType.name());
-    }
 
-    @Override
-    public void onPlaybackError(ErrorType errorType, String errorDetails) {
-        Log.d("MainActivity", "Playback error received: " + errorType.name());
-    }
-
-    @Override
-    protected void onDestroy() {
-        Spotify.destroyPlayer(this);
-        super.onDestroy();
-    }
 
 
     class RetrieveMySongsTask extends AsyncTask<Void, Void, String> {
