+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,10 @@
 <resources>
-    <string name="app_name">Music Driven Backyard Lightshow</string>
+    <string name="app_name">MBLS</string>
  <!--old stuff -->
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
-
+    <string name="spotify_main_menu">Spotify Home</string>
+    <string name="home">Home</string>
 
     <string name="app_name_short">kMP</string>
