+++ local.properties
@@ -7,6 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Tue Nov 24 17:46:08 EST 2015
-ndk.dir=C\:\\Users\\yvinogradov\\AppData\\Local\\Android\\sdk\\ndk-bundle
-sdk.dir=C\:\\Users\\yvinogradov\\AppData\\Local\\Android\\sdk
+#Wed Nov 25 21:14:07 EST 2015
+sdk.dir=C\:\\Users\\Joey\\AppData\\Local\\Android\\sdk
